refactor(problem4): 검증 로직을 별도의 클래스로 분리
